+++ CHANGES.md
@@ -1,6 +1,8 @@
 ## Version 2.0.0.dev9 (in dev)
 
+* Representative default variables [#590](https://github.com/CCI-Tools/cate/issues/590).
 * Tasks are no longer executed in parallel [#606](https://github.com/CCI-Tools/cate/issues/606).
+* WebAPI service problem in CLI [#600](https://github.com/CCI-Tools/cate/issues/600)
 * Improve error messages and handling [#393](https://github.com/CCI-Tools/cate/issues/393),
   introduced new error type `cate.core.types.ValidationError` for special treatment in the GUI.
 * Make Cate HTTP User-Agent distinguishable [#510](https://github.com/CCI-Tools/cate/issues/510).